% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyser_sentiment.R
\name{analyser-sentiment}
\alias{analyser-sentiment}
\alias{analyser_sentiment}
\alias{analyser_sentiments}
\title{Analyse de sentiment}
\usage{
analyser_sentiment(texte, inclure_tokens = FALSE)

analyser_sentiments(texte, retour = c("tout", "score", "vote"))
}
\arguments{
\item{texte}{Texte(s) sur lequel effectuer l'analyse de sentiment.}

\item{inclure_tokens}{Inclure les tokens extrait de \code{texte} dans la sortie.}

\item{retour}{Valeur de sortie : \code{"tout"} (un \code{data.frame} avec une ligne par \code{texte}),
\code{"score"} (score obtenu par texte), \code{"vote"} (positive ou negative).}
}
\value{
A \code{data.frame} with sentiment measure.
}
\description{
Analyse de sentiment
}
\examples{
analyser_sentiment("J'aime beaucoup la choucroute !")
}
