% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviner_langue.R
\name{deviner_langue}
\alias{deviner_langue}
\title{Deviner la langue}
\usage{
deviner_langue(texte)
}
\arguments{
\item{texte}{Une chaîne de caractère(s).}
}
\value{
Un \code{data.frame} de toutes les langues classées par ordre décroissant de leur score.
}
\description{
Devenier la langue utilisé dans un texte.
}
\examples{
deviner_langue("Bonjour, une baguette pas trop cuite, s'il vous plaît !")
}
