% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements.R
\name{normaliser}
\alias{normaliser}
\title{Normaliser un texte}
\usage{
normaliser(texte)
}
\arguments{
\item{texte}{Une chaîne de caractère(s).}
}
\value{
Une chaîne de caractère(s) normalisée.
}
\description{
Normaliser un texte
}
\examples{
normaliser("Bonjour, comment allez-vous ?")
normaliser(c("Bonjour, comment allez-vous ?", "Salut ! Yes impeccable"))
}
