% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements.R
\name{stemmatiser}
\alias{stemmatiser}
\title{Stemmatiser des mots}
\usage{
stemmatiser(mots)
}
\arguments{
\item{mots}{Un vecteur de mots.}
}
\value{
Un vecteur de mots stemmatisés.
}
\description{
Stemmatiser des mots
}
\examples{
stemmatiser(c("les", "chats", "ont", "faim"))
}
