% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements.R
\name{tokeniser}
\alias{tokeniser}
\title{Tokeniser un texte}
\usage{
tokeniser(texte, normaliser = TRUE)
}
\arguments{
\item{texte}{Une chaîne de caractère(s).}

\item{normaliser}{Normaliser le texte, voir \code{\link[=normaliser]{normaliser()}}.}
}
\value{
Une liste de tokens.
}
\description{
Tokeniser un texte
}
\examples{
tokeniser(c("Bonjour, comment allez-vous ?", "Salut ! Yes impeccable"))
}
