% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements.R
\name{traiter_texte}
\alias{traiter_texte}
\title{Traiter du texte}
\usage{
traiter_texte(texte, enlever_stopwords = TRUE)
}
\arguments{
\item{texte}{Une chaîne de caractère(s).}

\item{enlever_stopwords}{Enlever ou pas les mots stop.}
}
\value{
Une liste contenant les mots normalisés, tokenisés et stemmatisés.
}
\description{
Permet de \code{\link[=normaliser]{normaliser()}}, \code{\link[=tokeniser]{tokeniser()}} et \code{\link[=stemmatiser]{stemmatiser()}} du texte.
}
\examples{
traiter_texte(c("Les chats ont faim.", "Qui a laissé sortir les chiens ?"))
}
